Move test for sparse DataFrame to test_random_over_sampler.py
